Avoid stopping released CFRunLoop
As fsevents is saving the result of `CFRunLoopGetCurrent` and stopping
it from a separate thread, it's necessary to increase the reference
count of the returned CFRunLoopRef.

In the event that the run loop is terminated by other means,
`CFRunLoopStop` may be passed a deallocated reference and result in
a panic (SIGSEGV).